core/util: Working on Sean's comments
1) Added licenses to new files
2) Delete comments in ofi_mem_hooks.h
3) Renamed files
4) Dropped 'linux' prefix in names functions/vars
5) Took the ofi_get_page_size() func in libfaric src
6) Threw away SIZE_OF_VOID_P
7) Threw away while() and make a label for unlock in patch_symbol()
8) Removed _t prefix
9) Removed typedefs: But I got some warnings, while it is compiling.
